+++ clients/clients.go
@@ -29,27 +29,3 @@ func DefaultNewRequest(req_type, url string, data []byte) (req *http.Request, er
 	return req, err
 
 }
-
-type Context struct {
-	Components map[string]Component `json:"contextData"`
-}
-
-func NewContext() *Context {
-	return &Context{
-		Components: make(map[string]Component),
-	}
-}
-
-func (c *Context) AddComponent(key string, component Component) {
-	c.Components[key] = component
-}
-
-type Component struct {
-	Type string `json:"componentType"`
-}
-
-func NewComponent(t string) Component {
-	return Component{
-		Type: t,
-	}
-}
