+++ components/components.go
@@ -1,0 +1,11 @@
+package components
+
+type Component struct {
+	Type string `json:"componentType"`
+}
+
+func NewComponent(t string) Component {
+	return Component{
+		Type: t,
+	}
+}
